+++ app/components/RefreshBanner/messages.js
@@ -1,0 +1,13 @@
+/*
+ * RefreshBanner Messages
+ *
+ * This contains all the text for the RefreshBanner component.
+ */
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  notification: {
+    id: 'app.components.RefreshBanner.notification',
+    defaultMessage: 'There is a new version available! Click to refresh.',
+  },
+});