+++ app/containers/AgentDesktop/actions.js
@@ -57,6 +57,7 @@ import {
   UPDATE_SCRIPT_VALUES,
   SET_DISPOSITION_DETAILS,
   SELECT_DISPOSITION,
+  SHOW_REFRESH_NOTIF,
 } from './constants';
 
 export function setExtensions(response) {
@@ -441,5 +442,12 @@ export function selectDisposition(interactionId, disposition) {
     type: SELECT_DISPOSITION,
     interactionId,
     disposition,
+  };
+}
+
+export function showRefreshRequired(show) {
+  return {
+    type: SHOW_REFRESH_NOTIF,
+    show,
   };
 }